# openapi.yaml
openapi: 3.0.3
info:
  title: Doctor.mx API
  description: |
    Doctor.mx - Plataforma integral de telemedicina para México.
    
    Esta API proporciona acceso a todas las funcionalidades de la plataforma Doctor.mx,
    incluyendo consultas médicas virtuales, gestión de citas, recetas digitales,
    análisis de IA, consentimientos digitales, y más.
    
    ## Autenticación
    La API utiliza Supabase Auth para la autenticación. Incluye el token de sesión
    en el header `Authorization: Bearer <token>`.
    
    ## Roles
    - **patient**: Pacientes que buscan atención médica
    - **doctor**: Médicos que brindan consultas
    - **admin**: Administradores del sistema
    
    ## Rate Limiting
    Las solicitudes están limitadas por IP y usuario autenticado.
    Headers de respuesta:
    - `X-RateLimit-Limit`: Límite de solicitudes
    - `X-RateLimit-Remaining`: Solicitudes restantes
    - `X-RateLimit-Reset`: Tiempo de reinicio
  version: 1.0.0
  contact:
    name: Doctor.mx Support
    email: soporte@doctor.mx
  license:
    name: Propietario
    url: https://doctor.mx/terms
servers:
  - url: https://doctor.mx
    description: Producción
  - url: https://staging.doctor.mx
    description: Staging
  - url: http://localhost:3000
    description: Desarrollo local
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/doctor/analytics/ai-referrals:
    get:
      summary: Analytics de referencias IA
      description: Obtiene estadísticas de referencias generadas por IA
      tags:
        - Doctores
        - Analytics
      responses:
        '200':
          description: Estadísticas de referencias
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIReferralAnalytics'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/prescriptions/send:
    post:
      summary: Enviar múltiples
      description: Envía múltiples recetas
      tags:
        - Prescriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prescriptionIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Recetas enviadas

  # ==================== CHAT ====================
...
  /api/notifications/email/reminder:
    post:
      summary: Recordatorio
      description: Envía recordatorio de cita
      tags:
        - Notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                appointmentId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Recordatorio enviado

  # ==================== ANONYMOUS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de sesión de Supabase
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Cookie de sesión de Supabase
  schemas:
...
    AIReferralAnalytics:
      type: object
      properties:
        total:
          type: integer
        avgScore:
          type: number
        avgDoctorsAvailable:
          type: number
        avgDoctorsMatched:
          type: number
        bySpecialty:
          type: object
          additionalProperties:
            type: object
            properties:
              count:
                type: integer
              avgScore:
                type: number
        byUrgency:
          type: object
          additionalProperties:
            type: object
            properties:
              count:
                type: integer
              avgScore:
                type: number

    # ==================== SOAP NOTES ====================
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
...
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            message: "Se requiere autenticación"
...